exchange/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
urllib3==2.2.1
